+++ ipc/MetaFiles.ts
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
+import { z } from 'zod';
+
 export type UrlString = string;
 
 export interface SourceJson {
@@ -16,6 +18,18 @@ export type WithdrawnJson = UrlString[];
 export type AppVersions = {
     [version: string]: AppVersion;
 };
+
+type SemverString = string;
+
+export type NrfutilModuleName = string;
+export type NrfutilModuleVersion = SemverString;
+
+export interface NrfutilModules {
+    [name: NrfutilModuleName]: [
+        NrfutilModuleVersion,
+        ...NrfutilModuleVersion[]
+    ];
+}
 
 export type AppVersion = {
     shasum?: string;
@@ -38,48 +52,33 @@ export interface AppInfo {
     };
 }
 
-interface ObjectContainingOptionalStrings {
-    [index: string]: string | undefined;
-}
+export const PackageJsonSchema = z.object({
+    name: z.string(),
+    version: z.string(),
 
-interface NrfConnectForDesktop {
-    nrfutil?: NrfutilModules;
-    html?: string;
-}
+    author: z.string().optional(),
+    bin: z.union([z.record(z.string().optional()), z.string()]).optional(),
+    description: z.string(),
+    homepage: z.string(),
+    devDependencies: z.record(z.string().optional()).optional(),
+    displayName: z.string(),
+    engines: z.object({
+        nrfconnect: z.string(),
+    }),
+    nrfConnectForDesktop: z.object({
+        html: z.string().optional(),
+        nrfutil: z.record(z.array(z.string()).nonempty()).optional(),
+    }),
+    files: z.array(z.string()).nonempty().readonly(),
+    license: z.string().optional(),
+    main: z.string(),
+    dependencies: z.record(z.string().optional()).optional(),
+    peerDependencies: z.record(z.string().optional()).optional(),
+    repository: z.object({
+        type: z.literal('git'),
+        url: z.string(),
+    }),
+    scripts: z.record(z.string().optional()).optional(),
+});
 
-type SemverString = string;
-
-export type NrfutilModuleName = string;
-export type NrfutilModuleVersion = SemverString;
-
-export interface NrfutilModules {
-    [name: NrfutilModuleName]: [
-        NrfutilModuleVersion,
-        ...NrfutilModuleVersion[]
-    ];
-}
-
-export interface PackageJson {
-    name: string;
-    version: SemverString;
-
-    // Several optional properties
-    author?: string;
-    bin?: ObjectContainingOptionalStrings | string;
-    dependencies?: ObjectContainingOptionalStrings;
-    description?: string;
-    homepage?: UrlString;
-    devDependencies?: ObjectContainingOptionalStrings;
-    displayName?: string;
-    engines?: ObjectContainingOptionalStrings;
-    nrfConnectForDesktop?: NrfConnectForDesktop;
-    files?: readonly string[];
-    license?: string;
-    main?: string;
-    peerDependencies?: ObjectContainingOptionalStrings;
-    repository?: {
-        type: string;
-        url: UrlString;
-    };
-    scripts?: ObjectContainingOptionalStrings;
-}
+export type PackageJson = z.infer<typeof PackageJsonSchema>;